Move block sync test into alphanet. Make it pass. More formatting
