+++ node/alphanet/src/lib.rs
@@ -14,6 +14,7 @@ use futures::sync::mpsc::{channel, Receiver};
 
 use client::Client;
 use configs::{ClientConfig, get_testnet_configs, NetworkConfig, RPCConfig};
+use coroutines::importer::spawn_block_importer;
 use coroutines::ns_control_builder::get_control;
 use coroutines::ns_producer::spawn_block_producer;
 use network::spawn_network;
@@ -51,8 +52,9 @@ pub fn start_from_client(
         spawn_receipt_task(client.clone(), receipts_rx);
 
         // Launch block syncing / importing.
-        let (inc_block_tx, _inc_block_rx) = channel(1024);
+        let (inc_block_tx, inc_block_rx) = channel(1024);
         let (_out_block_tx, out_block_rx) = channel(1024);
+        spawn_block_importer(client.clone(), inc_block_rx);
 
         // Launch Nightshade task
         let (inc_gossip_tx, inc_gossip_rx) = channel(1024);
@@ -108,8 +110,11 @@ fn spawn_receipt_task(client: Arc<Client>, receipt_rx: Receiver<ReceiptBlock>) {
 
 #[cfg(test)]
 mod tests {
-    use primitives::transaction::TransactionBody;
+    use client::BlockProductionResult;
     use primitives::block_traits::SignedBlock;
+    use primitives::block_traits::SignedBlock;
+    use primitives::chain::ChainPayload;
+    use primitives::transaction::TransactionBody;
 
     use crate::testing_utils::{configure_chain_spec, Node, wait};
 
@@ -137,7 +142,8 @@ mod tests {
         );
 
         alice.client.shard_client.pool.add_transaction(
-            TransactionBody::send_money(1, "alice.near", "bob.near", 10).sign(&alice.secret_key));
+            TransactionBody::send_money(1, "alice.near", "bob.near", 10).sign(&alice.secret_key),
+        );
 
         alice.start();
         bob.start();
@@ -151,5 +157,27 @@ mod tests {
             500,
             10000,
         );
+    }
+
+
+    #[test]
+    fn test_two_nodes_sync() {
+        let chain_spec = configure_chain_spec();
+        let alice = Node::new("t2_alice", "alice.near", 1, "127.0.0.1:3002", 3032, vec![], chain_spec.clone());
+        let bob = Node::new("t2_bob", "bob.near", 2, "127.0.0.1:3003", 3033, vec![alice.node_info.clone()], chain_spec);
+
+        let payload = ChainPayload { transactions: vec![], receipts: vec![] };
+        let (beacon_block, shard_block) = match alice.client.try_produce_block(1, payload) {
+            BlockProductionResult::Success(beacon_block, shard_block) => (beacon_block, shard_block),
+            _ => panic!("Should produce block"),
+        };
+        alice.client.try_import_blocks(beacon_block, shard_block);
+
+        alice.start();
+        bob.start();
+
+        wait(|| {
+            bob.client.shard_client.chain.best_block().index() == 1
+        }, 500, 10000);
     }
 }
