+++ node/consensus/src/passthrough.rs
@@ -1,8 +1,8 @@
 use std::sync::Arc;
 use std::time::Duration;
 
-use futures::{future, Future, Sink, Stream};
 use futures::sync::mpsc::{Receiver, Sender};
+use futures::{future, Future, Sink, Stream};
 use tokio::{self, timer::Interval};
 
 use client::Client;
@@ -18,22 +18,25 @@ pub fn spawn_consensus(
 ) {
     let initial_beacon_block_index = client.beacon_chain.chain.best_index();
     let task = Interval::new_interval(block_period)
-        .fold((control_rx, initial_beacon_block_index), move |(control_rx, mut beacon_block_index), _| {
-            let hash = client.shard_client.pool.snapshot_payload();
-            if hash != CryptoHash::default() {
-                beacon_block_index += 1;
-                let c = ConsensusBlockHeader {
-                    header: BlockHeader { author: 0, hash },
-                    index: beacon_block_index
-                };
-                tokio::spawn(consensus_tx.clone().send(c).map(|_| ()).map_err(|e| {
-                    error!("Failure sending pass-through consensus {}", e);
-                }));
-                future::ok((control_rx, beacon_block_index))
-            } else {
-                future::ok((control_rx, beacon_block_index))
-            }
-        })
+        .fold(
+            (control_rx, initial_beacon_block_index),
+            move |(control_rx, mut beacon_block_index), _| {
+                let hash = client.shard_client.pool.snapshot_payload();
+                if hash != CryptoHash::default() {
+                    beacon_block_index += 1;
+                    let c = ConsensusBlockHeader {
+                        header: BlockHeader { author: 0, hash },
+                        index: beacon_block_index,
+                    };
+                    tokio::spawn(consensus_tx.clone().send(c).map(|_| ()).map_err(|e| {
+                        error!("Failure sending pass-through consensus {}", e);
+                    }));
+                    future::ok((control_rx, beacon_block_index))
+                } else {
+                    future::ok((control_rx, beacon_block_index))
+                }
+            },
+        )
         .map(|_| ())
         .map_err(|e| error!("timer error: {}", e));
