+++ tests/test_testnet.rs
@@ -6,9 +6,6 @@ use std::panic;
 use std::process::Command;
 
 use alphanet::testing_utils::{check_result, configure_chain_spec, Node, wait};
-use client::BlockProductionResult;
-use primitives::chain::ChainPayload;
-use primitives::block_traits::SignedBlock;
 
 #[test]
 #[ignore]
@@ -46,26 +43,4 @@ fn test_two_nodes() {
         check_result(res).is_ok()
     };
     wait(view_account, 500, 60000);
-}
-
-#[test]
-#[ignore]
-fn test_two_nodes_sync() {
-    let chain_spec = configure_chain_spec();
-    let alice = Node::new("t2_alice", "alice.near", 1, "127.0.0.1:3002", 3032, vec![], chain_spec.clone());
-    let bob = Node::new("t2_bob", "bob.near", 2, "127.0.0.1:3003", 3033, vec![alice.node_info.clone()], chain_spec);
-
-    let payload = ChainPayload { transactions: vec![], receipts: vec![] };
-    let (beacon_block, shard_block) = match alice.client.try_produce_block(1, payload) {
-        BlockProductionResult::Success(beacon_block, shard_block) => (beacon_block, shard_block),
-        _ => panic!("Should produce block"),
-    };
-    alice.client.try_import_blocks(beacon_block, shard_block);
-
-    alice.start();
-    bob.start();
-
-    wait(|| {
-        bob.client.shard_client.chain.best_block().index() == 1
-    }, 500, 10000);
 }
